backend/requirements.txt
# FinTerminal backend dependencies
# Install: pip install -r requirements.txt

# API framework
fastapi==0.111.0
uvicorn==0.29.0

# Data sources
yfinance>=1.2.0
fredapi==0.5.2
requests==2.31.0
pandas==2.2.2
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
apscheduler==3.10.4
httpx>=0.27.0
websockets>=12.0

# Auth (Supabase JWT verification)
PyJWT==2.9.0
